requirements.txt
dj-database-url==0.5.0
Django==2.0.7
django-easy-pdf==0.1.1
django-heroku==0.3.1
django-nocaptcha-recaptcha==0.0.20
django-widget-tweaks==1.4.2
gunicorn==19.9.0
html5lib==1.0.1
httplib2==0.11.3
Pillow==5.2.0
psycopg2==2.7.5
PyPDF2==1.26.0
python-decouple==3.1
pytz==2018.5
reportlab==3.5.2
six==1.11.0
webencodings==0.5.1
whitenoise==3.3.1
xhtml2pdf==0.2.2
